+++ packages/ethers/src/utils/defaultConfig.ts
@@ -1,6 +1,6 @@
 import '@web3modal/polyfills'
-import type { Metadata, Provider, ProviderType } from '@web3modal/scaffold-utils/ethers'
-import { CoinbaseWalletSDK } from '@coinbase/wallet-sdk'
+import type { Chain, Metadata, Provider, ProviderType } from '@web3modal/scaffold-utils/ethers'
+import { CoinbaseWalletSDK, type ProviderInterface } from '@coinbase/wallet-sdk'
 import type { SocialProvider } from '@web3modal/scaffold-utils'
 
 export interface ConfigOptions {
@@ -15,6 +15,8 @@ export interface ConfigOptions {
   rpcUrl?: string
   defaultChainId?: number
   metadata: Metadata
+  chains?: Chain[]
+  coinbasePreference?: 'all' | 'smartWalletOnly' | 'eoaOnly'
 }
 
 export function defaultConfig(options: ConfigOptions) {
@@ -32,7 +34,7 @@ export function defaultConfig(options: ConfigOptions) {
   let injectedProvider: Provider | undefined = undefined
 
   // eslint-disable-next-line @typescript-eslint/no-redundant-type-constituents
-  let coinbaseProvider: Provider | undefined = undefined
+  let coinbaseProvider: ProviderInterface | undefined = undefined
 
   const providers: ProviderType = { metadata }
 
@@ -67,11 +69,20 @@ export function defaultConfig(options: ConfigOptions) {
     const coinbaseWallet = new CoinbaseWalletSDK({
       appName: metadata.name,
       appLogoUrl: metadata.icons[0],
-      darkMode: false,
-      enableMobileWalletLink: true
+      appChainIds: options.chains?.map(chain => chain.chainId) || [1, 84532]
     })
 
-    coinbaseProvider = coinbaseWallet.makeWeb3Provider(rpcUrl, defaultChainId)
+    /**
+     * Determines which wallet options to display in Coinbase Wallet SDK.
+     * @property options
+     *   - `all`: Show both smart wallet and EOA options.
+     *   - `smartWalletOnly`: Show only smart wallet options.
+     *   - `eoaOnly`: Show only EOA options.
+     * @see https://www.smartwallet.dev/sdk/v3-to-v4-changes#parameters
+     */
+    coinbaseProvider = coinbaseWallet.makeWeb3Provider({
+      options: options.coinbasePreference || 'all'
+    })
 
     return coinbaseProvider
   }