+++ packages/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @web3modal/common
 
+## 4.2.3
+
+### Patch Changes
+
+- feat: - feat: restricted ens names. fix: iat set automatically if not present in messageParams. Adds siwe config handlers
+
+## 4.2.3-alpha.0
+
+### Patch Changes
+
+- feat: add support for coinbase smart accounts
+
 ## 4.2.2
 
 ### Patch Changes